Fixed Textures, Added Mouse Input/nLooking to move all input tto SDL_Event, to allow for mousewheel use

+++ src/myengine/Window.h
@@ -26,8 +26,8 @@ namespace myEngine {
       Window(std::string _name, int _width, int _height);
       ~Window();
 
-			float getWidth() { return (float)m_width; }
-			float getHeight() { return (float)m_height; }
+			int getWidth() { return m_width; }
+			int getHeight() { return m_height; }
 
 	};
